Improve Playwright performance and fix warnings
